Rename and namespace AudioComponent.
